+++ NEWS.md
@@ -4,6 +4,13 @@ This page describes the most important changes in `EasyFITS`. The format is base
 Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic
 Versioning](https://semver.org).
 
+## Unreleased
+
+### Fixed
+
+- `writefits!` always overwrite even though `overwrite=false` is specified.
+
+
 ## Version 0.7.2 [2026-01-08]
 
 ### Fixed